move a azure shared code test file to non-azure
